+++ src/postgres.rs
@@ -1,5 +1,5 @@
 pub use deadpool_postgres;
-use std::{future::Future, str::FromStr, time::Duration};
+use std::{future::Future, time::Duration};
 
 use chrono::prelude::*;
 use cron::Schedule;
@@ -11,7 +11,7 @@ use crate::{execute_sleep, LoopState};
 pub fn make_looper<Fut1, Fut2>(
     pg_pool: deadpool_postgres::Pool,
     token: CancellationToken,
-    expression: &str,
+    schedule: Schedule,
     stop_check_duration: Duration,
     task_function: impl Fn(DateTime<Utc>, Result<deadpool_postgres::Client, deadpool_postgres::PoolError>) -> Fut1
         + Send
@@ -26,10 +26,14 @@ where
     Fut1: Future<Output = LoopState> + Send,
     Fut2: Future<Output = ()> + Send,
 {
-    let expression = expression.to_owned();
     spawn(async move {
-        let schedule = Schedule::from_str(&expression).unwrap();
-        let mut next_tick: DateTime<Utc> = schedule.upcoming(Utc).next().unwrap();
+        let mut next_tick: DateTime<Utc> = match schedule.upcoming(Utc).next() {
+            Some(next_tick) => next_tick,
+            None => {
+                stop_function(pg_pool.get().await).await;
+                return;
+            }
+        };
         loop {
             // グレースフルストップのチェック
             if token.is_cancelled() {
@@ -46,6 +50,7 @@ where
                 {
                     next_tick = res;
                 } else {
+                    stop_function(pg_pool.get().await).await;
                     break;
                 }
             }
@@ -92,6 +97,7 @@ where
                 {
                     next_tick = res;
                 } else {
+                    stop_function(pg_pool.get().await).await;
                     break;
                 }
             }