+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 ## Changes
 
+### v0.1.1 (2024/02/27)
+* remove unnecessary features
+* remove unnecessary unwrap
+* cron crate reexport
+* fix: don't call stop_fuction in terminate state
+
 ### v0.1.0 (2024/02/22)
 * first release